ajout d'un son d'ambiance en jeu
ajout d'un son quand le joueur se déplace.
Les bactéries ne générent pas bc de déchets mais les autres unités oui.

+++ InsideInvaders/Assets/Components/Absorbeur.cs
@@ -3,5 +3,5 @@
 public class Absorbeur : MonoBehaviour {
 	// Advice: FYFY component aims to contain only public members (according to Entity-Component-System paradigm).
 	public float rayon_effet = 30;
-	public int degats_absorbtion = 1;
+	public float degats_absorbtion = 1;
 }
