+++ InsideInvaders/Assets/Components/Agglutineur.cs
@@ -3,5 +3,6 @@
 public class Agglutineur : MonoBehaviour {
 	// Advice: FYFY component aims to contain only public members (according to Entity-Component-System paradigm).
 	public float rayon_effet = 30;
-	public int vitesse_agglutinement = 1;
+	//public int vitesse_agglutinement = 1;
+	public float vitesse_agglutinement = 1;
 }
