+++ InsideInvaders/Assets/Components/Infectieux.cs
@@ -3,5 +3,5 @@
 public class Infectieux : MonoBehaviour {
 	// Advice: FYFY component aims to contain only public members (according to Entity-Component-System paradigm).
 	public float rayon_effet = 30;
-	public int vitesse_infection = 1;
+	public float vitesse_infection = 1;
 }
